♻️ Move traits and related impls to their parent module directly
