+++ src/service.rs
@@ -1,9 +1,3 @@
 //! Handle externals like OS, API, or DB
 pub mod archive;
 pub mod fs;
-pub mod os;
-
-pub use os::LocalFileSystemService;
-
-#[cfg(test)]
-pub use os::MockFileSystemService;
